Use top-level slugify websocket command
